Add kafka-stat-lag module to recover topics-partitions lag stats.
